refactor(graphql): deprecate the ontology query and move it in a new file
US-123 US-124
